Update degrees per pixel calculation
For really small FOV:s, the assumption that the hud is spherical works
well. But that's not true for normal FOV:s

therefore, do the calculation for degrees per pixel at center of screen
so the horizon line sticks to the environment when at level.

Improves the "stickiness" of the HUD to the environment